rutils.iter: fix defclause-sequence drivers
Defaults `step` in `return-seq-code` to 1.

Currently, all drivers that use `defclause-sequence` are broken, because
`step` is not bound before `return-sequence-code` is called. This
results in `step` being bound to nil, which means `initial-value` is
set to `(- nil)` and `step-code` is set to something like `(+ index
nil)`. This change causes `step` to default to 1 if it is not provided.